+++ simple-stack-example-mvp/src/main/java/com/zhuinden/simplestackdemoexamplemvp/presentation/paths/first/FirstView.java
@@ -2,7 +2,6 @@ package com.zhuinden.simplestackdemoexamplemvp.presentation.paths.first;
 
 import android.annotation.TargetApi;
 import android.content.Context;
-import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.util.AttributeSet;
 import android.widget.RelativeLayout;
@@ -10,6 +9,7 @@ import android.widget.RelativeLayout;
 import com.squareup.coordinators.Coordinators;
 import com.zhuinden.simplestack.Backstack;
 import com.zhuinden.simplestack.Bundleable;
+import com.zhuinden.simplestack.StateBundle;
 import com.zhuinden.simplestackdemoexamplemvp.application.CustomApplication;
 
 import javax.inject.Inject;
@@ -57,13 +57,13 @@ public class FirstView
     }
 
     @Override
-    public Bundle toBundle() {
+    public StateBundle toBundle() {
         FirstCoordinator firstCoordinator = Coordinators.getCoordinator(this);
         return firstCoordinator.toBundle();
     }
 
     @Override
-    public void fromBundle(@Nullable Bundle bundle) {
+    public void fromBundle(@Nullable StateBundle bundle) {
         FirstCoordinator firstCoordinator = Coordinators.getCoordinator(this);
         firstCoordinator.fromBundle(bundle);
     }